+++ app/Jobs/Notifications/BeatmapsetDiscussionUnlock.php
@@ -3,8 +3,8 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Models\Notifications;
+namespace App\Jobs\Notifications;
 
-class BeatmapsetDiscussionLock extends BeatmapsetNotification
+class BeatmapsetDiscussionUnlock extends BeatmapsetNotification
 {
 }
