+++ app/Jobs/Notifications/BeatmapsetNotification.php
@@ -3,14 +3,14 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Models\Notifications;
+namespace App\Jobs\Notifications;
 
 use App\Models\BeatmapDiscussionPost;
 use App\Models\Beatmapset;
 use App\Models\User;
 use App\Models\UserNotificationOption;
 
-abstract class BeatmapsetNotification extends NotificationBase
+abstract class BeatmapsetNotification extends BroadcastNotification
 {
     protected static function beatmapsetWatcherUserIds($beatmapset)
     {
@@ -20,7 +20,7 @@ abstract class BeatmapsetNotification extends NotificationBase
         );
     }
 
-    public function __construct($object, ?User $source)
+    public function __construct($object, ?User $source = null)
     {
         parent::__construct($object, $source);
