+++ app/Jobs/Notifications/ForumTopicReply.php
@@ -3,13 +3,13 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace App\Models\Notifications;
+namespace App\Jobs\Notifications;
 
 use App\Models\Forum\Post;
 use App\Models\User;
 use App\Models\UserNotificationOption;
 
-class ForumTopicReply extends NotificationBase
+class ForumTopicReply extends BroadcastNotification
 {
     public function __construct(Post $object, ?User $source)
     {
