--- app/Models/Notifications/NotificationBase.php
@@ -1,89 +1,0 @@
-<?php
-
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
-// See the LICENCE file in the repository root for full licence text.
-
-namespace App\Models\Notifications;
-
-use App\Models\Notification;
-use App\Models\User;
-use App\Models\UserNotificationOption;
-
-abstract class NotificationBase
-{
-    const CONTENT_TRUNCATE = 36;
-
-    protected $notifiable;
-    protected $notificationName;
-    protected $object;
-    protected $source;
-
-    public static function filterUserIdsForNotificationOption(array $userIds, $optionName)
-    {
-        // FIXME: filtering all the ids could get quite large?
-        $notificationOptions = UserNotificationOption
-            ::whereIn('user_id', $userIds)
-            ->where(['name' => $optionName])
-            ->whereNotNull('details')
-            ->get()
-            ->keyBy('user_id');
-
-        $filteredUserIds = [];
-        foreach ($userIds as $userId) {
-            if ($notificationOptions[$userId]->details['push'] ?? true) {
-                $filteredUserIds[] = $userId;
-            }
-        }
-
-        return $filteredUserIds;
-    }
-
-    public static function notificationClassFor(string $name)
-    {
-        return get_class_namespace(static::class).'\\'.studly_case($name);
-    }
-
-    public function __construct($object, ?User $source)
-    {
-        $this->object = $object;
-        $this->source = $source;
-
-        $this->notificationName = snake_case(get_class_basename(get_class($this)));
-    }
-
-    public function makeNotification(): Notification
-    {
-        $params['details'] = $this->getDetails();
-        $params['name'] = $this->notificationName;
-        if ($this->source !== null) {
-            $params['details']['username'] = $this->source;
-        }
-
-        $notification = new Notification($params);
-        $notification->notifiable()->associate($this->getNotifiable());
-        if ($this->source !== null) {
-            $notification->source()->associate($this->source);
-        }
-
-        return $notification;
-    }
-
-    abstract function getDetails(): array;
-
-    abstract function getListentingUserIds(): array;
-
-    public function getNotifiable()
-    {
-        return $this->notifiable ?? $this->object;
-    }
-
-    public function getReceiverIds(): array
-    {
-        return array_values(array_unique(array_diff($this->getListentingUserIds(), [optional($this->source)->getKey()])));
-    }
-
-    public function getSource(): ?User
-    {
-        return $this->source;
-    }
-}
