+++ tests/Controllers/BeatmapDiscussionPostsControllerTest.php
@@ -6,7 +6,9 @@
 namespace Tests\Controllers;
 
 use App\Events\NewPrivateNotificationEvent;
-use App\Jobs\BroadcastNotification;
+use App\Jobs\Notifications\BeatmapsetDiscussionPostNew;
+use App\Jobs\Notifications\BeatmapsetDiscussionQualifiedProblem;
+use App\Jobs\Notifications\BeatmapsetDisqualify;
 use App\Models\Beatmap;
 use App\Models\BeatmapDiscussion;
 use App\Models\BeatmapDiscussionPost;
@@ -752,23 +754,13 @@ class BeatmapDiscussionPostsControllerTest extends TestCase
                 ],
             ]);
 
-        $remaining = $queued;
-        // assertPushed only asserts if any matching job was queued.
-        Queue::assertPushed(BroadcastNotification::class, function ($job) use ($queued, &$remaining) {
-            $inArray = in_array($job->getName(), $queued, true);
+        foreach ($queued as $class) {
+            Queue::assertPushed($class);
+        }
 
-            if (($key = array_search($job->getName(), $remaining, true)) !== false) {
-                unset($remaining[$key]);
-            }
-
-            return $inArray;
-        });
-
-        $this->assertEmpty(array_values($remaining));
-
-        Queue::assertNotPushed(BroadcastNotification::class, function ($job) use ($notQueued) {
-            return in_array($job->getName(), $notQueued, true);
-        });
+        foreach ($notQueued as $class) {
+            Queue::assertNotPushed($class);
+        }
     }
 
     public function problemDataProvider()
@@ -784,13 +776,13 @@ class BeatmapDiscussionPostsControllerTest extends TestCase
         return [
             [
                 'bng',
-                [Notification::BEATMAPSET_DISQUALIFY, Notification::BEATMAPSET_DISCUSSION_POST_NEW],
-                [Notification::BEATMAPSET_DISCUSSION_QUALIFIED_PROBLEM],
+                [BeatmapsetDisqualify::class, BeatmapsetDiscussionPostNew::class],
+                [BeatmapsetDiscussionQualifiedProblem::class],
             ],
             [
                 null,
-                [Notification::BEATMAPSET_DISCUSSION_POST_NEW, Notification::BEATMAPSET_DISCUSSION_QUALIFIED_PROBLEM],
-                [Notification::BEATMAPSET_DISQUALIFY],
+                [BeatmapsetDiscussionPostNew::class, BeatmapsetDiscussionQualifiedProblem::class],
+                [BeatmapsetDisqualify::class],
             ],
         ];
     }
